refactor(cms): remove cards collection and use reusable content instead
closed COD-322
