+++ libs/app-cms/util/plugins/src/lib/plugins/get-multi-tenant-plugin.ts
@@ -12,7 +12,6 @@ export const getMultiTenantPlugin = () =>
       name: 'tenant'
     },
     collections: {
-      cards: {},
       categories: {},
       forms: {},
       'form-submissions': {},
