+++ libs/shared/util/payload-types/src/lib/custom-types.ts
@@ -2,6 +2,7 @@ import type { StripTypes } from '@codeware/shared/util/typesafe';
 import type { ClientUser, User } from 'payload';
 
 import type {
+  CardBlock,
   User as CollectionUser,
   Config,
   ContentBlock,
@@ -49,6 +50,9 @@ export type UserAny = ClientUser | CollectionUser | User;
 
 /** Block type */
 export type BlockSlug = keyof Config['blocks'];
+
+/** Card block card details */
+export type CardBlockCard = NonNullable<CardBlock['cards']>[number];
 
 /** Content block column size */
 export type ContentBlockSize = NonNullable<
